Fix build on FreeBSD
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@796956 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/thread/src/thread_native_basic.c
@@ -888,7 +888,7 @@ IDATA VMCALL hythread_set_state(hythread_t thread, IDATA state) {
 } // hythread_set_state
 
 IDATA VMCALL hythread_get_thread_id_offset() {
-    return (U_32)&((HyThread *)0)->thread_id;
+    return (U_32)(size_t)&((HyThread *)0)->thread_id;
 } // hythread_get_thread_id_offset
 
 IDATA VMCALL hythread_set_thread_stop_callback(hythread_t thread,
